Revert "Much Java optimization"
This reverts commit 7aee576b2401dcd51e8d370b4ccfd86dc982fd7c.

+++ jdosbox-win/src/main/java/jdos/win/kernel/VideoMemory.java
@@ -15,7 +15,7 @@ public class VideoMemory {
         int frame = WinSystem.memory.frames.length*32;
         long address = WinSystem.getCurrentProcess().addressSpace.getNextAddress(WinProcess.ADDRESS_VIDEO_START, amount, true);
         WinSystem.getCurrentProcess().addressSpace.alloc(address, amount);
-        frame += (int) ((address- WinProcess.ADDRESS_VIDEO_START + 0xFFF) >>> 12);
+        frame += (address- WinProcess.ADDRESS_VIDEO_START + 0xFFF) >>> 12;
         int directory = WinSystem.getCurrentProcess().page_directory;
         for (long i=address;i<address+amount;i+=0x1000) {
             int pagePtr = WinSystem.memory.get_page((int)i, true, directory);
