+++ jdosbox/src/main/java/jdos/util/IntPtr.java
@@ -24,10 +24,10 @@ public class IntPtr extends Ptr {
         return 4;
     }
     public void set(int off, int val) {
-        writed(off, val);
+        writed((int)off, val);
     }
     public int get(int off) {
-        return readd(off);
+        return (int)readd(off);
     }
     public void set(int off, long val) {
         writed(off, val);
